+++ packages/backend/src/models/entities/meta.ts
@@ -454,4 +454,9 @@ export class Meta {
 		default: false,
 	})
 	public enableIpLogging: boolean;
+
+	@Column('boolean', {
+		default: true,
+	})
+	public enableActiveEmailValidation: boolean;
 }